+++ src/main/java/io/github/astrapi69/mystic/crypt/panel/dbtree/SecretKeyTreeWithContentPanel.java
@@ -26,6 +26,8 @@ package io.github.astrapi69.mystic.crypt.panel.dbtree;
 
 import java.awt.event.MouseEvent;
 import java.io.Serial;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -36,13 +38,22 @@ import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.DefaultTreeModel;
 import javax.swing.tree.TreeModel;
 
+import net.miginfocom.layout.AC;
+import net.miginfocom.layout.CC;
+import net.miginfocom.layout.LC;
+import net.miginfocom.swing.MigLayout;
+
 import org.jdesktop.swingx.JXTree;
 
+import io.github.astrapi69.browser.BrowserControlExtensions;
+import io.github.astrapi69.clone.CloneQuietlyExtensions;
 import io.github.astrapi69.design.pattern.observer.event.EventObject;
 import io.github.astrapi69.design.pattern.observer.event.EventSource;
 import io.github.astrapi69.gen.tree.BaseTreeNode;
 import io.github.astrapi69.gen.tree.TreeIdNode;
 import io.github.astrapi69.gen.tree.convert.BaseTreeNodeTransformer;
+import io.github.astrapi69.gen.tree.visitor.MaxIndexFinderTreeNodeVisitor;
+import io.github.astrapi69.gen.tree.visitor.ReindexTreeNodeVisitor;
 import io.github.astrapi69.id.generate.LongIdGenerator;
 import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.IModel;
@@ -63,6 +74,7 @@ import io.github.astrapi69.swing.tree.JTreeExtensions;
 import io.github.astrapi69.swing.tree.panel.content.BaseTreeNodeGenericTreeElementWithContentPanel;
 import io.github.astrapi69.swing.tree.panel.node.NodePanel;
 import io.github.astrapi69.swing.tree.renderer.state.NewGenericBaseTreeNodeCellRenderer;
+import io.github.astrapi69.swing.utils.ClipboardExtensions;
 
 public class SecretKeyTreeWithContentPanel
 	extends
@@ -208,14 +220,19 @@ public class SecretKeyTreeWithContentPanel
 					actionEvent -> this.onAddNewChildTreeNode(mouseEvent)));
 			}
 
+			popup.add(JMenuItemFactory.newJMenuItem("Edit node...",
+				actionEvent -> this.onEditSelectedTreeNode(mouseEvent)));
+
+
 			if (!selectedTreeNodeElement.isRoot())
 			{
+				popup.add(JMenuItemFactory.newJMenuItem("Duplicate node...",
+					actionEvent -> this.onDuplicateSelectedTreeNode(mouseEvent)));
+
 				popup.add(JMenuItemFactory.newJMenuItem("delete",
 					actionEvent -> this.onDeleteSelectedTreeNode(mouseEvent)));
 			}
 
-			popup.add(JMenuItemFactory.newJMenuItem("Edit node...",
-				actionEvent -> this.onEditSelectedTreeNode(mouseEvent)));
 
 			popup.add(JMenuItemFactory.newJMenuItem("Collapse node",
 				actionEvent -> this.onCollapseSelectedTreeNode(mouseEvent)));
@@ -228,14 +245,103 @@ public class SecretKeyTreeWithContentPanel
 	}
 
 	/**
+	 * The callback method on duplicate a tree node
+	 */
+	@SuppressWarnings("unchecked")
+	protected void onDuplicateSelectedTreeNode(MouseEvent mouseEvent)
+	{
+		JTreeExtensions.getSelectedDefaultMutableTreeNode(mouseEvent, tree)
+			.ifPresent(selectedDefaultMutableTreeNode -> {
+				BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long> selectedTreeNode = (BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long>)selectedDefaultMutableTreeNode
+					.getUserObject();
+				BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long> parentTreeNode = selectedTreeNode
+					.getParent();
+				ReindexTreeNodeVisitor<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long, BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long>> reindexTreeNodeVisitor;
+				MaxIndexFinderTreeNodeVisitor<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long, BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long>> maxIndexFinderTreeNodeVisitor;
+
+				Long maxIndex;
+				Long nextId;
+				maxIndexFinderTreeNodeVisitor = new MaxIndexFinderTreeNodeVisitor<>()
+				{
+					@Override
+					public boolean isGreater(Long id)
+					{
+						return getMaxIndex() < id;
+					}
+				};
+				BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long> clonedTreeNode = CloneQuietlyExtensions
+					.clone(selectedTreeNode);
+				NodePanel panel = new NodePanel()
+				{
+					protected void onInitializeMigLayout()
+					{
+						MigLayout layout = new MigLayout(new LC().fillX().wrapAfter(2),
+							new AC().align("left").gap("10").grow().fill(),
+							new AC().fill().gap("10"));
+						this.setLayout(layout);
+
+						add(getLblName());
+						add(getTxtName(), new CC().grow().width("120px"));
+					}
+				};
+				String newName = clonedTreeNode.getDisplayValue() + "-Copy";
+				panel.getModelObject().setName(newName);
+				panel.getTxtName().setText(newName);
+				int option = JOptionPaneExtensions.getSelectedOption(panel,
+					JOptionPane.PLAIN_MESSAGE, JOptionPane.OK_CANCEL_OPTION, null,
+					Messages.getString("dialog.duplicate.node.entry.title", "Name for duplicate"),
+					panel.getTxtName());
+				if (option == JOptionPane.OK_OPTION)
+				{
+					NodeModel modelObject = panel.getModelObject();
+					newName = modelObject.getName();
+
+					clonedTreeNode.getValue().setName(newName);
+					clonedTreeNode.setDisplayValue(newName);
+					clonedTreeNode.setParent(parentTreeNode);
+
+					parentTreeNode.addChild(clonedTreeNode);
+
+					selectedTreeNode.getRoot().accept(maxIndexFinderTreeNodeVisitor);
+					maxIndex = maxIndexFinderTreeNodeVisitor.getMaxIndex();
+
+					nextId = maxIndex + 1;
+					MysticCryptApplicationFrame.getInstance()
+						.setIdGenerator(LongIdGenerator.of(nextId));
+					LongIdGenerator idGenerator = MysticCryptApplicationFrame.getInstance()
+						.getIdGenerator();
+					reindexTreeNodeVisitor = new ReindexTreeNodeVisitor<>(idGenerator);
+					clonedTreeNode.accept(reindexTreeNodeVisitor);
+
+					BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long> rootTreeNode = selectedTreeNode
+						.getRoot();
+					Map<Long, TreeIdNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long>> clonedKeyMap = BaseTreeNodeTransformer
+						.toKeyMap(clonedTreeNode);
+					Map<Long, TreeIdNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long>> longTreeIdNodeMap = BaseTreeNodeTransformer
+						.toKeyMap(rootTreeNode);
+					longTreeIdNodeMap.putAll(clonedKeyMap);
+					MysticCryptApplicationFrame.getInstance().getModelObject()
+						.setRootTreeAsMap(longTreeIdNodeMap);
+
+					DefaultMutableTreeNode parent = (DefaultMutableTreeNode)selectedDefaultMutableTreeNode
+						.getParent();
+
+					BaseTreeNodeFactory.newDefaultMutableTreeNode(clonedTreeNode, parent, false);
+
+					reload(parent);
+				}
+			});
+	}
+
+	/**
 	 * The callback method on add a new child tree node
 	 */
 	@SuppressWarnings("unchecked")
 	protected void onAddNewChildTreeNode(MouseEvent mouseEvent)
 	{
 		JTreeExtensions.getSelectedDefaultMutableTreeNode(mouseEvent, tree)
-			.ifPresent(selectedTreeNode -> {
-				Object userObject = selectedTreeNode.getUserObject();
+			.ifPresent(selectedDefaultMutableTreeNode -> {
+				Object userObject = selectedDefaultMutableTreeNode.getUserObject();
 				BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long> parentTreeNode = (BaseTreeNode<GenericTreeElement<List<MysticCryptEntryModelBean>>, Long>)userObject;
 				NodePanel panel = new NodePanel();
 				int option = JOptionPaneExtensions.getSelectedOption(panel,
@@ -261,8 +367,8 @@ public class SecretKeyTreeWithContentPanel
 					parentTreeNode.addChild(newTreeNode);
 
 					DefaultMutableTreeNode newChild = new DefaultMutableTreeNode(newTreeNode, leaf);
-					selectedTreeNode.add(newChild);
-					reload(selectedTreeNode);
+					selectedDefaultMutableTreeNode.add(newChild);
+					reload(selectedDefaultMutableTreeNode);
 				}
 			});
 	}
@@ -376,20 +482,81 @@ public class SecretKeyTreeWithContentPanel
 
 		JPopupMenu popup = JPopupMenuFactory.newJPopupMenu();
 
-		popup.add(JMenuItemFactory.newJMenuItem("add ...", actionEvent -> this.onAddTableEntry()));
+		JMenuItem copyUsername = JMenuItemFactory.newJMenuItem("Copy Password",
+			actionEvent -> this.onCopyUsernameTableEntry());
+		copyUsername.setEnabled(allSelectedRowData.size() == 1);
+		popup.add(copyUsername);
 
-		JMenuItem menuItem = JMenuItemFactory.newJMenuItem("delete",
-			actionEvent -> this.onDeleteTableEntry());
-		menuItem.setEnabled(!allSelectedRowData.isEmpty());
-		popup.add(menuItem);
+		JMenuItem copyPassword = JMenuItemFactory.newJMenuItem("Copy Password",
+			actionEvent -> this.onCopyPasswordTableEntry());
+		copyPassword.setEnabled(allSelectedRowData.size() == 1);
+		popup.add(copyPassword);
+
+		JMenuItem openUrl = JMenuItemFactory.newJMenuItem("Open url",
+			actionEvent -> this.onOpenUrlOfTableEntry());
+		openUrl.setEnabled(
+			allSelectedRowData.size() == 1 && allSelectedRowData.get(0).getUrl() != null);
+		popup.add(openUrl);
+
+		// Separator
+		popup.addSeparator();
+
+		JMenuItem add = JMenuItemFactory.newJMenuItem("add ...",
+			actionEvent -> this.onAddTableEntry());
+		popup.add(add);
 
 		JMenuItem edit = JMenuItemFactory.newJMenuItem("edit",
 			actionEvent -> this.onEditTableEntry());
 		edit.setEnabled(allSelectedRowData.size() == 1);
 		popup.add(edit);
 
-		popup.show(getTblTreeEntryTable(), x, y);
+		JMenuItem duplicate = JMenuItemFactory.newJMenuItem("duplicate",
+			actionEvent -> this.onDuplicateTableEntry());
+		duplicate.setEnabled(allSelectedRowData.size() == 1);
+		popup.add(duplicate);
 
+		JMenuItem delete = JMenuItemFactory.newJMenuItem("delete",
+			actionEvent -> this.onDeleteTableEntry());
+		delete.setEnabled(!allSelectedRowData.isEmpty());
+		popup.add(delete);
+		// Separator
+		popup.addSeparator();
+
+		JMenuItem selectAll = JMenuItemFactory.newJMenuItem("select all",
+			actionEvent -> this.onSelectAllTableEntries());
+		selectAll.setEnabled(0 < getTblTreeEntryTable().getRowCount());
+		popup.add(selectAll);
+
+		popup.show(getTblTreeEntryTable(), x, y);
+	}
+
+	protected void onSelectAllTableEntries()
+	{
+		getTblTreeEntryTable().selectAll();
+	}
+
+	protected void onDuplicateTableEntry()
+	{
+		getTblTreeEntryTable().getSingleSelectedRowData().ifPresent(tableEntry -> {
+
+			// TODO...
+		});
+	}
+
+	protected void onCopyUsernameTableEntry()
+	{
+		getTblTreeEntryTable().getSingleSelectedRowData().ifPresent(tableEntry -> {
+			String userName = tableEntry.getUserName();
+			ClipboardExtensions.copyToClipboard(userName);
+		});
+	}
+
+	protected void onCopyPasswordTableEntry()
+	{
+		getTblTreeEntryTable().getSingleSelectedRowData().ifPresent(tableEntry -> {
+			char[] password = tableEntry.getPassword();
+			ClipboardExtensions.copyToClipboard(String.valueOf(password));
+		});
 	}
 
 	protected void onDeleteTableEntry()
@@ -412,6 +579,23 @@ public class SecretKeyTreeWithContentPanel
 	{
 		getTblTreeEntryTable().getSingleSelectedRowData().ifPresent(tableEntry -> {
 			showEditMysticCryptEntryDialog(tableEntry);
+		});
+	}
+
+	protected void onOpenUrlOfTableEntry()
+	{
+		getTblTreeEntryTable().getSingleSelectedRowData().ifPresent(tableEntry -> {
+			String urlString = tableEntry.getUrl();
+			try
+			{
+				URL url = new URL(urlString);
+				url.toExternalForm();
+				BrowserControlExtensions.displayURLonStandardBrowser(this, urlString);
+			}
+			catch (MalformedURLException e)
+			{
+				throw new RuntimeException(e);
+			}
 		});
 	}