+++ src/test/java/io/github/astrapi69/mystic/crypt/panel/dbtree/NewMysticCryptEntryFormPanel.java
@@ -1,3 +1,27 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
@@ -10,6 +34,25 @@ package io.github.astrapi69.mystic.crypt.panel.dbtree;
  */
 public class NewMysticCryptEntryFormPanel extends javax.swing.JPanel
 {
+
+	// Variables declaration - do not modify//GEN-BEGIN:variables
+	private javax.swing.JButton btnGeneratePassword;
+	private javax.swing.JButton btnShowPassword;
+	private javax.swing.JCheckBox cbxExpirable;
+	private javax.swing.JLabel lblEntryName;
+	private javax.swing.JLabel lblNotes;
+	private javax.swing.JLabel lblPassword;
+	private javax.swing.JLabel lblRepeat;
+	private javax.swing.JLabel lblUrl;
+	private javax.swing.JLabel lblUsername;
+	private javax.swing.JScrollPane srcNotes;
+	private javax.swing.JTextField txtEntryName;
+	private javax.swing.JTextField txtExpires;
+	private javax.swing.JTextArea txtNotes;
+	private javax.swing.JPasswordField txtPassword;
+	private javax.swing.JPasswordField txtRepeat;
+	private javax.swing.JTextField txtUrl;
+	private javax.swing.JTextField txtUsername;
 
 	/**
 	 * Creates new form NewMysticCryptEntryFormPanel
@@ -190,25 +233,5 @@ public class NewMysticCryptEntryFormPanel extends javax.swing.JPanel
 	{// GEN-FIRST:event_txtEntryNameActionPerformed
 		// TODO add your handling code here:
 	}// GEN-LAST:event_txtEntryNameActionPerformed
-
-
-	// Variables declaration - do not modify//GEN-BEGIN:variables
-	private javax.swing.JButton btnGeneratePassword;
-	private javax.swing.JButton btnShowPassword;
-	private javax.swing.JCheckBox cbxExpirable;
-	private javax.swing.JLabel lblEntryName;
-	private javax.swing.JLabel lblNotes;
-	private javax.swing.JLabel lblPassword;
-	private javax.swing.JLabel lblRepeat;
-	private javax.swing.JLabel lblUrl;
-	private javax.swing.JLabel lblUsername;
-	private javax.swing.JScrollPane srcNotes;
-	private javax.swing.JTextField txtEntryName;
-	private javax.swing.JTextField txtExpires;
-	private javax.swing.JTextArea txtNotes;
-	private javax.swing.JPasswordField txtPassword;
-	private javax.swing.JPasswordField txtRepeat;
-	private javax.swing.JTextField txtUrl;
-	private javax.swing.JTextField txtUsername;
-	// End of variables declaration//GEN-END:variables
+		// End of variables declaration//GEN-END:variables
 }