+++ CHANGELOG.md
@@ -1,9 +1,28 @@
 ## Change log
 ----------------------
 
+Version 7.2
+-------------
+
+ADDED:
+
+- new popup menu for duplicate an existing tree node
+
+CHANGED:
+
+- removed spring-boot dependencies
+
 Version 7.1
 -------------
 
+ADDED:
+
+- new button for show the password in the mystic entry dialog
+- new button for generate a new password in the mystic entry dialog
+
+CHANGED:
+
+- the new dialog is synchronized now with the sign in dialog
 
 Version 7
 -------------